+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - Fixed the CSS for ProgressBar and added tooltip on hover for uploaded text in progressBar header.
 
 ### Fixed
+- Fixed Progress Bar Storybook update
 
 ### Changed